requirements.txt
cmsis-svd>=0.4
absl-py>=0.11.0